### 20240428, v0.8.3 Build 84 * include static code analysis configuration `toroCLibrary.ruleset` * fix minor static code analysis warnings

+++ toroCLibrary/LibCore/_cdeCoreWcsStrTok.c
@@ -72,7 +72,7 @@ void* _cdeCoreWcsStrTok(CDE_APP_IF* pCdeAppIf, IN void* pvoidStr, IN const void*
                 *((short*)pNew) = '\0';
             else
                 *((char*)pNew) = '\0';
-            *(char**)ppLast = (char*)pNew + (1 + pParm->fWide);
+            *(char**)ppLast = (char*)pNew + (size_t)(1 + pParm->fWide);
         }
         else {
             *ppLast = NULL;                                     // we have reasched the end
